+++ CHANGELOG.md
@@ -113,6 +113,9 @@ Increment the:
 * [SEMANTIC CONVENTIONS] Generate event name constants (e.g. `semconv::exception::kException`)
   [#4171](https://github.com/open-telemetry/opentelemetry-cpp/pull/4171)
 
+* [EXPORTER] Handle OTLP partial success response
+  [#4104](https://github.com/open-telemetry/opentelemetry-cpp/pull/4104)
+
 ## [1.27.0] 2026-05-13
 
 * [RELEASE] Bump main branch to 1.27.0-dev